Fin du cours 29 octobre - joyeux anniversaire

+++ app/src/main/java/ca/qc/cstj/spaceexplorer/models/Planet.kt
@@ -1,4 +1,4 @@
 package ca.qc.cstj.spaceexplorer.models
 
-data class Planet(val name:String, val temperature:Float)
+data class Planet(val name:String, val temperature:Double)
